start working on a region marker gui
Automatic spatial clustering is not possible because all the lateral horns look quite different. But there are still typical areas. So we decided to first do this by hand. The idea is to not do the ICA simultaneously as we did in the last weeks. We do ICA or NMF on a single animal. The result are eg. 5 spatial and 5 temporal modes. We extract the outlines of the spatial modes by morphological operators and plot them with different colors in one imagesc plot. On its right side we plot all temporal modes in one plot. The idea of the GUI is now to assign labels to each of the spatial mode outlines.