3d valenz scatterplot
same scatterplot as before, three dimensional points are placed according to how strong they activate a spatial mode (cluster). But this time they are colored by the valenz of an odor.

the colormap is jet and the valenz is normalized so that neutral valenz is green, negative blue and positive red. Odors for which we don't have valenz information are painted black.