investigate relation between activation of clusters and valenz
the question is which cluster or which combination of cluster is responsible for the aversion of certain odors. Is it one of the clusters we found, the ratio between activations or the difference between cluster activations.

the newly added party calculates correlations for these models

unnormalized valenz is used for this..